Switched link's  health to use double instead of int
AB@210

+++ LegendOfZelda/Interface/IPlayer.cs
@@ -2,7 +2,7 @@
 {
     public interface IPlayer : IDynamic
     {
-        void BeHealthy(int healAmount);
+        void BeHealthy(double healAmount);
         void BeDamaged(double damage);
         void MoveUp();
         void MoveDown();
