+++ LegendOfZelda/Link/State/LinkActiveAbstractState.cs
@@ -46,7 +46,7 @@ namespace LegendOfZelda.Link.State
 
         public abstract void UseBomb();
 
-        public void BeDamaged(int damage)
+        public void BeDamaged(double damage)
         {
             if (!damaged)
             {
@@ -56,7 +56,7 @@ namespace LegendOfZelda.Link.State
             }
         }
 
-        public void BeHealthy(int healAmount)
+        public void BeHealthy(double healAmount)
         {
             damaged = false;
             link.AddHealth(healAmount);
